requirements.txt
apache-airflow
fairlearn
dvc[gs]
scikit-learn
pandas
numpy
google-cloud-storage
google-auth
matplotlib
seaborn
hyperopt
mlflow
pytest
pandera
xgboost


